requirements.txt
matrix-nio[e2e]==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
pylast==5.2.0
aiosqlite==0.19.0
python-dateutil==2.8.2
